+++ Makefile.autosetup
@@ -361,6 +361,7 @@ LIBDEBUGOBJS+=	debug/parse_test.o
 LIBDEBUGOBJS+=	debug/window.o
 @endif
 @if HAVE_LIBUNWIND || USE_DEBUG_COLOR || USE_DEBUG_EMAIL || USE_DEBUG_GRAPHVIZ || USE_DEBUG_NOTIFY || USE_DEBUG_PARSE_TEST || USE_DEBUG_WINDOW
+LIBDEBUGOBJS+=	debug/common.o
 LIBDEBUG=	libdebug.a
 CLEANFILES+=	$(LIBDEBUG) $(LIBDEBUGOBJS)
 ALLOBJS+=	$(LIBDEBUGOBJS)
