+++ debug/lib.h
@@ -28,6 +28,7 @@
  * | File                | Description                |
  * | :------------------ | :------------------------- |
  * | debug/backtrace.c   | @subpage debug_backtrace   |
+ * | debug/common.c      | @subpage debug_common      |
  * | debug/email.c       | @subpage debug_email       |
  * | debug/graphviz.c    | @subpage debug_graphviz    |
  * | debug/notify.c      | @subpage debug_notify      |
@@ -52,6 +53,12 @@ struct ListHead;
 struct MailboxView;
 struct NotifyCallback;
 struct PagerPrivateData;
+
+// Common
+void        add_flag               (struct Buffer *buf, bool is_set, const char *name);
+const char *get_content_encoding   (enum ContentEncoding enc);
+const char *get_content_disposition(enum ContentDisposition disp);
+const char *get_content_type       (enum ContentType type);
 
 // Backtrace
 void show_backtrace(void);
